requirements.txt
greenlet
gevent
geventhttpclient
gunicorn
pycassa
python-cjson
werkzeug
requests>=1.2.0
